REF: Move `OCTUE_SERVICES_NAMESPACE` to package root
