+++ octue/cloud/pub_sub/subscription.py
@@ -13,7 +13,7 @@ from google.pubsub_v1.types.pubsub import (
     UpdateSubscriptionRequest,
 )
 
-from octue.cloud.service_id import OCTUE_SERVICES_NAMESPACE
+from octue import OCTUE_SERVICES_NAMESPACE
 from octue.exceptions import ConflictingSubscriptionType
